# faster hex-prefix codec and string -> []byte