--- a/trie/encoding.go
+++ b/trie/encoding.go
@@ -18,11 +18,9 @@ package trie
 
 import (
 	"bytes"
-	"encoding/hex"
-	"strings"
 )
 
-func CompactEncode(hexSlice []byte) string {
+func CompactEncode(hexSlice []byte) []byte {
 	terminator := 0
 	if hexSlice[len(hexSlice)-1] == 16 {
 		terminator = 1
@@ -45,10 +43,10 @@ func CompactEncode(hexSlice []byte) string {
 		buff.WriteByte(byte(16*hexSlice[i] + hexSlice[i+1]))
 	}
 
-	return buff.String()
+	return buff.Bytes()
 }
 
-func CompactDecode(str string) []byte {
+func CompactDecode(str []byte) []byte {
 	base := CompactHexDecode(str)
 	base = base[:len(base)-1]
 	if base[0] >= 2 {
@@ -63,30 +61,23 @@ func CompactDecode(str string) []byte {
 	return base
 }
 
-func CompactHexDecode(str string) []byte {
-	base := "0123456789abcdef"
-	var hexSlice []byte
+func CompactHexDecode(str []byte) []byte {
+	var nibbles []byte
 
-	enc := hex.EncodeToString([]byte(str))
-	for _, v := range enc {
-		hexSlice = append(hexSlice, byte(strings.IndexByte(base, byte(v))))
+	for _, b := range str {
+		nibbles = append(nibbles, b/16)
+		nibbles = append(nibbles, b%16)
 	}
-	hexSlice = append(hexSlice, 16)
-
-	return hexSlice
+	nibbles = append(nibbles, 16)
+	return nibbles
 }
 
-func DecodeCompact(key []byte) string {
-	const base = "0123456789abcdef"
-	var str string
-
-	for _, v := range key {
-		if v < 16 {
-			str += string(base[v])
-		}
+// assumes key is odd length
+func DecodeCompact(key []byte) []byte {
+	var res []byte
+	for i := 0; i < len(key)-1; i += 2 {
+		v1, v0 := key[i], key[i+1]
+		res = append(res, v1*16+v0)
 	}
-
-	res, _ := hex.DecodeString(str)
-
-	return string(res)
+	return res
 }