Some improvements to the hex-prefix encoder/decoder:
- use []byte for all function arguments and returns (rather than strings, which are just being converted to []byte everywhere)
- use integer based encoder/decoder rather than string concatenation and pushing through a hex codec

Benchmarks:

Before the changes:
----------------------------
BatBuddha:trie$ go test --bench=Get
PASS
BenchmarkGets    1000000              1696 ns/op
ok      github.com/ethereum/go-ethereum/trie    1.913s


With this PR:
----------------
BatBuddha:trie$ go test --bench=Get
PASS
BenchmarkGets    2000000               966 ns/op
ok      github.com/ethereum/go-ethereum/trie    3.125s